Extract project-related execution services
